+++ src/arguments/resources/index.ts
@@ -3,3 +3,4 @@ export * from './lootTables'
 export * from './predicate'
 export * from './recipe'
 export * from './tag'
+export * from './itemModifier'
