+++ src/datapack/BasePath.ts
@@ -1,10 +1,11 @@
 import {
-  AdvancementInstance, LootTableInstance, MCFunctionClass, PredicateInstance, RecipeInstance, TagInstance,
+  AdvancementInstance, ItemModifierInstance, ItemModifierOptions, LootTableInstance, MCFunctionClass, PredicateInstance, RecipeInstance, TagInstance,
 } from '@resources'
 
 import type {
   AdvancementJSON, HintedTagStringType, LootTableJSON, PredicateJSON, RecipeJSON, TAG_TYPES,
   TagSingleValue,
+  ItemModifierJSON,
 } from 'src/arguments'
 import type { HideFunctionProperties } from '@/generalTypes'
 import type { Datapack } from '@datapack'
@@ -64,6 +65,7 @@ export class BasePathClass<N extends (undefined | string) = (undefined | string)
       onConflict.predicate ??= onConflict.default
       onConflict.recipe ??= onConflict.default
       onConflict.tag ??= onConflict.default
+      onConflict.itemModifier ??= onConflict.default
     }
 
     this.onConflict = onConflict
@@ -265,4 +267,9 @@ export class BasePathClass<N extends (undefined | string) = (undefined | string)
   Tag = <T extends TAG_TYPES>(type: T, name: string, values: TagSingleValue<HintedTagStringType<T>>[] = [], replace?: boolean, options?: TagOptions) => (
     new TagInstance(this.datapack, type, this.getName(name), values, replace, { onConflict: this.onConflict?.tag, ...options })
   )
+
+  /** Create an item modifier */
+  ItemModifier = (name: string, modifier: ItemModifierJSON, options?: ItemModifierOptions) => {
+    new ItemModifierInstance(this.datapack, this.getName(name), modifier, { onConflict: this.onConflict?.predicate, ...options })
+  }
 }
