+++ src/index.ts
@@ -6,7 +6,7 @@ export const {
 } = dataPack
 
 export const {
-  MCFunction, Advancement, Predicate, Tag, LootTable, Recipe,
+  MCFunction, Advancement, Predicate, Tag, LootTable, Recipe, ItemModifier
 } = dataPack.basePath
 
 export { _ } from './init'
@@ -40,5 +40,5 @@ export * from './datapack/BasePath'
 export * from './datapack/Datapack'
 export type { LiteralUnion } from './generalTypes'
 export type {
-  AdvancementInstance, LootTableInstance, MCFunctionOptions, PredicateInstance, RecipeInstance, ResourceInstance, TagInstance,
+  AdvancementInstance, LootTableInstance, MCFunctionOptions, PredicateInstance, RecipeInstance, ResourceInstance, TagInstance, ItemModifierInstance
 } from './resources'
