+++ src/resources/index.ts
@@ -5,3 +5,4 @@ export * from './Predicate'
 export * from './Recipe'
 export * from './Resource'
 export * from './Tag'
+export * from './ItemModifier'
